services/frontend/requirements.txt
streamlit
streamlit-agraph
# neo4j==6.1.0
langchain==1.2.7
langchain-neo4j==0.8.0
# langchain-community==0.4.1
langchain-ollama==1.0.1
langchain-openai
python-dotenv